==> 1. High frequency submission..ipynb <==
# Submission solely based on frequency of the product overall.

You process only data from `2016-05-28`

Get the frequency of products on 2016-05-28.
and for each user find the product that is the most frequent from above.. that is not present already.

import pandas


header = ["ncodpers","ind_ahor_fin_ult1","ind_aval_fin_ult1","ind_cco_fin_ult1","ind_cder_fin_ult1","ind_cno_fin_ult1","ind_ctju_fin_ult1","ind_ctma_fin_ult1","ind_ctop_fin_ult1","ind_ctpp_fin_ult1","ind_deco_fin_ult1","ind_deme_fin_ult1","ind_dela_fin_ult1","ind_ecue_fin_ult1","ind_fond_fin_ult1","ind_hip_fin_ult1","ind_plan_fin_ult1","ind_pres_fin_ult1","ind_reca_fin_ult1","ind_tjcr_fin_ult1","ind_valo_fin_ult1","ind_viv_fin_ult1","ind_nomina_ult1","ind_nom_pens_ult1","ind_recibo_ult1"]

tr = pandas.read_csv('data/train_2016_05_28.csv', header=None, names=header, usecols=[1] + range(24, 48))

freq = []
for col in header[1:]:
    freq.append((col, sum(tr[col])))

freq = sorted(freq, cmp = lambda x, y: cmp(x[1], y[1]), reverse = True)
sorted_products = map(lambda x: x[0], freq)

print freq[:5]
print sorted_products[:5]

te = pandas.read_csv('data/test_ver2.csv', usecols = [1])

te.head()

new_tr = tr.set_index(["ncodpers"], drop = True)

x = new_tr.loc[15892]

x['ind_ahor_fin_ult1']

te["added_products"] = ["ind_tjcr_fin_ult1"]*te.shape[0]

te.head()

for i in range(50): #te.index:
    if i % 100000 == 0:
        print i, "rows processed"
    person = te.loc[i]["ncodpers"]
    #print person
    row = new_tr.loc[person]
    #print row
    added_products = ""
    temp = 0
    for product in sorted_products:
        #print row[product], product
        if row[product] != 1:
            added_products += product + " "
            #print added_products
            temp += 1
            if temp == 7:
                break
    added_products.strip()
    #print added_products
    te.set_value(i,"added_products",added_products)

te.head()

te.to_csv('data/submissions/top_frequent_unused_products.csv', index=False)

from scripts.kaggle.helpers import make_submission
make_submission('data/submissions/top_frequent_unused_products.csv', description="get all the unsubscribed products that from high frequency to low frequency", submit=False, compress=True)

make_submission?
