==> 2. Based on simple independent Bayesian Probabilities..ipynb <==
# Based on simple bayesian probabilities.

The idea is that given previous months products that a user is subscribed to,  
identify new products that he will be interested in.


lets just consider 2016-05 data.. and from identify what products a given  
user will be interested in 2016-06

for example consider two products A, B  
calculate P(B = 1 on 05th given A = 0 on 04th) and P(B = 1 on 05th given A = 1 on 04th)  
While calculating these probabilities you don't consider data where B = 1 on 04th


Month |  A  |  B
----|----|----
2016-04 | 0 | 0
2016-05 | 1 | 1


from collections import Counter

n = 24 # number of products

counters = [[Counter() for _ in range(n)] for i in range(n)]
old = [0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1]
new = [1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1]

def updateCounter(old, new, counters):
    """
    given older months product data and 
    newer months product data update counter object
    old = [0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1]
    new = [1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1]
    
    counters[i][j] contains counter, c = {'00': m, '01': n, '11': o, '10': p}
    and p(j = 1 on new given i = 1 on old) = (c['11']*1.0)/(c['10'] + c['11'])
    and p(j = 1 on new given i = 0 on old) = (c['01']*1.0)/(c['00'] + c['01]))
    """
    n = len(old)
    for i in range(n):
        for j in range(n):
            if old[j] == 1:
                # have this condition if a user subscribed to a product 
                # in the older month.
                continue
            a = int(old[i])
            b = int(new[j])
            ab = str(a) + str(b)
            counters[i][j][ab] += 1
            
    
    return counters

def updateCounter2(old, new, counters):
    """
    given older months product data and 
    newer months product data update counter object
    old = [0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 0, 1]
    new = [1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1, 0, 1, 1]
    
    counters[i][j] contains counter, c = {'00': m, '01': n, '11': o, '10': p}
    and p(j = 1 on new given i = 1 on old) = (c['11']*1.0)/(c['10'] + c['11'])
    and p(j = 1 on new given i = 0 on old) = (c['01']*1.0)/(c['00'] + c['01]))
    """
    n = len(old)
    for i in range(n):
        for j in range(n):
            #if old[j] == 1:
            #    # have this condition if a user subscribed to a product 
            #    # in the older month.
            #    continue
            a = int(old[i])
            b = int(new[j])
            ab = str(a) + str(b)
            counters[i][j][ab] += 1
            
    
    return counters

counters = [[Counter() for _ in range(4)] for i in range(4)]
old = [0, 1, 0, 1]
new = [1, 1, 0, 1]
newCounters = updateCounter(old, new, counters)

newCounters

import pandas
from tqdm import tqdm



header = ["ncodpers","ind_ahor_fin_ult1","ind_aval_fin_ult1","ind_cco_fin_ult1","ind_cder_fin_ult1","ind_cno_fin_ult1","ind_ctju_fin_ult1","ind_ctma_fin_ult1","ind_ctop_fin_ult1","ind_ctpp_fin_ult1","ind_deco_fin_ult1","ind_deme_fin_ult1","ind_dela_fin_ult1","ind_ecue_fin_ult1","ind_fond_fin_ult1","ind_hip_fin_ult1","ind_plan_fin_ult1","ind_pres_fin_ult1","ind_reca_fin_ult1","ind_tjcr_fin_ult1","ind_valo_fin_ult1","ind_viv_fin_ult1","ind_nomina_ult1","ind_nom_pens_ult1","ind_recibo_ult1"]

old_tr = pandas.read_csv('data/train_2016_04_28.csv', header=None, names=header, usecols=[1] + range(24, 48), index_col=0)
new_tr = pandas.read_csv('data/train_2016_05_28.csv', header=None, names=header, usecols=[1] + range(24, 48), index_col=0)

indices = list(new_tr.index)

n = 24 # number of products

counters = [[Counter() for _ in range(n)] for i in range(n)]
for i in tqdm(indices):
    try:
        old = list(old_tr.loc[i])
        new = list(new_tr.loc[i])
        counters = updateCounter(old, new, counters)
    except Exception:
        pass

counters2 = [[Counter() for _ in range(n)] for i in range(n)]

for i in tqdm(indices):
    try:
        old = list(old_tr.loc[i])
        new = list(new_tr.loc[i])
        counters2 = updateCounter2(old, new, counters2)
    except Exception:
        pass

counters2[0][1]

def get_prob_matrix(counters):
    """
    returns two prob matrices
    prob0[i][j] contains P(j_new = 1/i_old = 0)
    prob1[i][j] contains P(j_new = 1/i_old = 1)
    """
    prob0 = [[0.0 for i in range(n)] for j in range(n)]
    prob1 = [[0.0 for i in range(n)] for j in range(n)]
    for i in range(n):
        for j in range(n):
            c = counters[i][j]
            prob1[i][j] = ((c['11'] + 1)*1.0)/(c['10'] + c['11'] + 2)
            prob0[i][j] = ((c['01'] + 1)*1.0)/(c['00'] + c['01'] + 2)
    return (prob0, prob1)

from operator import mul

def get_probabilities_of_unsubscribed_prods(products, counters = counters):
    """
    products = subscription data [0, 1, 1, 0, 0, 1]
    returns a list of products in descending order of probs 
    """
    prob0, prob1 = get_prob_matrix(counters)
    product_names = header[1:]
    prob_dict = {}
    for i, prod in enumerate(products):
        if prod == 0:
            temp_list = []
            for j in range(n):
                if i == j:
                    continue
                if products[j] == 0:
                    temp_list.append(prob0[j][i])
                if products[i] == 1:
                    temp_list.append(prob1[i][j])
            prob_dict[product_names[i]] = reduce(mul, temp_list, 1)
    sorted_prods = sorted(prob_dict.items(), cmp = lambda x, y: cmp(x[1], y[1]), reverse = True)[:7]
    # print sorted_prods
    product_string = ' '.join(map(lambda x: x[0], sorted_prods))
    return product_string
    

get_probabilities_of_unsubscribed_prods([0]*24)

te = pandas.read_csv('data/test_ver2.csv', usecols = [1])
te["added_products"] = ["ind_tjcr_fin_ult1"]*te.shape[0]



for i in te.index:
    if i % 100000 == 0:
        print i, "rows processed"
    person = te.loc[i]["ncodpers"]
    # print i, person
    row = list(new_tr.loc[person])
    # print row
    added_products = get_probabilities_of_unsubscribed_prods(row)
    # print added_products
    te.set_value(i,"added_products",added_products)

te.to_csv('data/submissions/2_bayesian_prob_products_counters_with_old_1_ignored.csv', index=False)

from scripts.kaggle.helpers import make_submission
make_submission('data/submissions/2_bayesian_prob_products_counters_with_old_1_ignored.csv', description="2. based on bayesian probabilities, where probs are calculated only for prods when there is a transition from 0 to 1", submit=True, compress=True)

## Cross validation code

# get predictions for 2016-05-28 from 2016-04-28
te = pandas.DataFrame()
te['ncodpers'] = new_tr.index
te['added_products'] = ["ind_tjcr_fin_ult1"]*te.shape[0]

for i in te.index:
    if i % 100000 == 0:
        print i, "rows processed"
    person = te.loc[i]["ncodpers"]
    # print i, person
    try:
        row = list(old_tr.loc[person])
    except:
        row = [0] * 24
    # print row
    added_products = get_probabilities_of_unsubscribed_prods(row)
    # print added_products
    te.set_value(i,"added_products",added_products)

#actual = pandas.read_csv('data/cv_2016_05_28.csv', index_col = 0)

def apk(actual, predicted, k=7):
    if len(predicted) > k:
        predicted = predicted[:k]

    score = 0.0
    num_hits = 0.0

    for i, p in enumerate(predicted):
        if p in actual and p not in predicted[:i]:
            num_hits += 1.0
            score += num_hits / (i+1.0)

    if not actual:
        return 0.0

    return score / min(len(actual), k)

h = lambda x: apk(x[0].split(), x[1].split())

actual = pandas.read_csv('data/cv_2016_05_28.csv', index_col = 0)
actual['predicted'] = te["added_products_03_04_1"]

actual = actual.fillna('a')

actual['apk'] = actual.apply(h, axis = 1)

more_apk = actual[actual.iloc[:,2] > 0.5]

numpy.mean(actual['apk'])

# Cross validation results
Cross validation with probs gotten from 04 to 05 transition and applied on 04
mean apk = 0.52377206914256291

almost 40k predictions were exact in this way..

not a very good measure.. so calculating countres from the trassision from 03 to 04 and applied on 04 subscriptions

more_apk.head()



header = ["ncodpers","ind_ahor_fin_ult1","ind_aval_fin_ult1","ind_cco_fin_ult1","ind_cder_fin_ult1","ind_cno_fin_ult1","ind_ctju_fin_ult1","ind_ctma_fin_ult1","ind_ctop_fin_ult1","ind_ctpp_fin_ult1","ind_deco_fin_ult1","ind_deme_fin_ult1","ind_dela_fin_ult1","ind_ecue_fin_ult1","ind_fond_fin_ult1","ind_hip_fin_ult1","ind_plan_fin_ult1","ind_pres_fin_ult1","ind_reca_fin_ult1","ind_tjcr_fin_ult1","ind_valo_fin_ult1","ind_viv_fin_ult1","ind_nomina_ult1","ind_nom_pens_ult1","ind_recibo_ult1"]

old_tr = pandas.read_csv('data/train_2016_03_28.csv', header=None, names=header, usecols=[1] + range(24, 48), index_col=0)
new_tr = pandas.read_csv('data/train_2016_04_28.csv', header=None, names=header, usecols=[1] + range(24, 48), index_col=0)

indices = list(new_tr.index)

n = 24 # number of products

counters3 = [[Counter() for _ in range(n)] for i in range(n)]
for i in tqdm(indices):
    try:
        old = list(old_tr.loc[i])
        new = list(new_tr.loc[i])
        counters3 = updateCounter(old, new, counters3)
    except Exception:
        pass

te['added_products_03_04_1'] = ["ind_tjcr_fin_ult1"]*te.shape[0]

for i in te.index:
    if i % 100000 == 0:
        print i, "rows processed"
    person = te.loc[i]["ncodpers"]
    # print i, person
    try:
        row = list(old_tr.loc[person])
    except:
        row = [0] * 24
    # print row
    added_products = get_probabilities_of_unsubscribed_prods(row, counters = counters3)
    # print added_products
    te.set_value(i,"added_products_03_04_1",added_products)

del(actual)

actual = pandas.read_csv('data/cv_2016_05_28.csv', index_col = 0)
actual['predicted'] = te["added_products_03_04_1"]

actual = actual.fillna('a')

actual['apk'] = actual.apply(h, axis = 1)

numpy.mean(actual['apk'])

te.tail()

more_apk = actual[actual.iloc[:,2] > 0.5]

te.tail()

x = te[te.iloc[:, 1] == te.iloc[:, 2]]

x.shape

sum(old_tr[header[1]])

 map(lambda x: (sum(old_tr[x]) + 100.0)/old_tr.shape[0], header[1:])

old_tr.loc[1567]

from scripts import bayesian

td = bayesian.generate_submission((2016, 5, 28), (2016, 6, 28))

from scripts.kaggle.helpers import make_submission
filename = 'data/submissions/proper_bayesian_submission_user_defined_by_previous_subscriptions.csv'
description = 'bayes submission, where user is defined by his previous subscriptions'
test_data.to_csv(filename, index=False)
make_submission(filename, description=description, submit=True, compress=True)
# make_submission('data/submissions/2_bayesian_prob_products_counters_with_old_1_ignored.csv', description="2. based on bayesian probabilities, where probs are calculated only for prods when there is a transition from 0 to 1", submit=True, compress=True)
