==> 15. Jacobian, Cosine similarity features generation.ipynb <==
import pandas as pd
import numpy as np
import xgboost as xgb
from sklearn.preprocessing import LabelEncoder
from scripts.kaggle.helpers import make_submission
import datetime
from tqdm import tqdm
import gc
import itertools
from multiprocessing import Pool
import pickle
import collections
import math

%matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns

# %pylab inline
# pylab.rcParams['figure.figsize'] = (12, 16)
plt.figure(figsize=(12,16))

%load_ext autotime

HEADER = ["fecha_dato", "ncodpers", "ind_empleado",
          "pais_residencia", "sexo", "age", "fecha_alta",
          "ind_nuevo", "antiguedad", "indrel", "ult_fec_cli_1t",
          "indrel_1mes", "tiprel_1mes", "indresi", "indext",
          "conyuemp", "canal_entrada", "indfall", "tipodom",
          "cod_prov", "nomprov", "ind_actividad_cliente",
          "renta", "segmento", "ind_ahor_fin_ult1",
          "ind_aval_fin_ult1", "ind_cco_fin_ult1",
          "ind_cder_fin_ult1", "ind_cno_fin_ult1",
          "ind_ctju_fin_ult1", "ind_ctma_fin_ult1",
          "ind_ctop_fin_ult1", "ind_ctpp_fin_ult1",
          "ind_deco_fin_ult1", "ind_deme_fin_ult1",
          "ind_dela_fin_ult1", "ind_ecue_fin_ult1",
          "ind_fond_fin_ult1", "ind_hip_fin_ult1",
          "ind_plan_fin_ult1", "ind_pres_fin_ult1",
          "ind_reca_fin_ult1", "ind_tjcr_fin_ult1",
          "ind_valo_fin_ult1", "ind_viv_fin_ult1",
          "ind_nomina_ult1", "ind_nom_pens_ult1",
          "ind_recibo_ult1"]

def get_persons_of_feature(train):
    persons_of_feature = dict()
    for feat in HEADER[24:]:
        persons_of_feature[feat] = set(train.loc[train[feat] == 1]["ncodpers"])
    return persons_of_feature
# for feat in header[24:]:
#     print feat, len(persons_of_feature[feat])

def jacobian_similarity(persons_of_feature, featur1, featur2):
    feat1 = persons_of_feature[featur1]
    feat2 = persons_of_feature[featur2]
    len1 = len(feat1)
    len2 = len(feat2)
    intersection = len(feat1.intersection(feat2))
    union = len1 + len2 - intersection
    if len1 == 0 or len1 == 0:
        return 0.0
    return intersection * 1.0 / union

def cosine_similarity(persons_of_feature, featur1, featur2):
    feat1 = persons_of_feature[featur1]
    feat2 = persons_of_feature[featur2]
    len1 = len(feat1)
    len2 = len(feat2)
    intersection = len(feat1.intersection(feat2))
    # union = len1 + len2 - intersection
    if len1 == 0 or len2 == 0:
        return 0.0
    return intersection * 1.0 / math.sqrt(len1*len2)

def get_similarity_matrices(train_df):
    persons_of_feature = get_persons_of_feature(train_df)
    jacob_matrix = collections.defaultdict(dict)
    cosine_matrix = collections.defaultdict(dict)
    for i in range(24):
        jacob_matrix[i][i] = 1.0
        cosine_matrix[i][i] = 1.0
        for j in range(i + 1, 24):
            temp = jacobian_similarity(persons_of_feature, HEADER[i + 24], HEADER[j+24])
            jacob_matrix[i][j] = temp
            jacob_matrix[j][i] = temp

            temp1 = cosine_similarity(persons_of_feature, HEADER[i+24], HEADER[j+24])
            cosine_matrix[i][j] = temp1
            cosine_matrix[j][i] = temp1
    cosine_pd = pd.DataFrame(cosine_matrix)
    jacob_pd = pd.DataFrame(jacob_matrix)
    return jacob_pd, cosine_pd


def get_persons_of_feature_from_added(added):
    persons_of_feature = dict()
    for feat in HEADER[24:]:
        persons_of_feature[feat] = set(added.loc[added['added_product'] == feat]["ncodpers"])
    return persons_of_feature
# for feat in header[24:]:
#     print feat, len(persons_of_feature[feat])

def get_similarity_matrices_from_added(added_df):
    persons_of_feature = get_persons_of_feature_from_added(added_df)
    jacob_matrix = collections.defaultdict(dict)
    cosine_matrix = collections.defaultdict(dict)
    for i in range(24):
        jacob_matrix[i][i] = 1.0
        cosine_matrix[i][i] = 1.0
        for j in range(i + 1, 24):
            temp = jacobian_similarity(persons_of_feature, HEADER[i + 24], HEADER[j+24])
            jacob_matrix[i][j] = temp
            jacob_matrix[j][i] = temp

            temp1 = cosine_similarity(persons_of_feature, HEADER[i+24], HEADER[j+24])
            cosine_matrix[i][j] = temp1
            cosine_matrix[j][i] = temp1
    cosine_pd = pd.DataFrame(cosine_matrix)
    jacob_pd = pd.DataFrame(jacob_matrix)
    return jacob_pd, cosine_pd

def get_wts_from_mat(df, row):
    brow = map(bool, row)
    wts = df.loc[brow, :].apply(sum)
    wts[brow] = 0
    return wts

def get_similarity_features(file_names): #train_file, added_file):
    train_file = file_names[0]
    added_file = file_names[1]
    print added_file
    train = pd.read_csv(train_file, header=None, names=HEADER[1:2] + HEADER[24:], usecols=[1]+range(24,48))
    train.fillna(0, inplace=True)
    for col in HEADER[24:]:
        train[col] = train[col].astype(int)
    if added_file == 'data/test_ver2.csv':
        added_products = pd.read_csv(added_file, usecols=[1])
    else:
        added_products = pd.read_csv(added_file, usecols=[0])
    jacob_pd, cosine_pd = get_similarity_matrices(train)
    train.set_index('ncodpers', inplace=True)
    train_final = pd.DataFrame()
    train_final['ncodpers'] = added_products['ncodpers']
    train_final.set_index('ncodpers', inplace=True)
    for col in HEADER[24:]:
        train_final['jacob_' + col] = [0.0]*train_final.shape[0]
        train_final['cosine_' + col] = [0.0]*train_final.shape[0]

    for i in tqdm(train_final.index):
        row = list(train.loc[i, HEADER[24:]])
        jwts = get_wts_from_mat(jacob_pd, row)
        cwts = get_wts_from_mat(cosine_pd, row)
        train_final.loc[i, ['jacob_' + h for h in HEADER[24:]]] = list(jwts)
        train_final.loc[i, ['cosine_' + h for h in HEADER[24:]]] = list(cwts)
    return train_final, file_names[2]

def get_similarity_features_custom(file_names): #train_file, added_file):
    train_file = file_names[0]
    added_file = file_names[1]
    print added_file
    train = pd.read_csv(train_file, header=None, names=HEADER[1:2] + HEADER[23:], usecols=[1]+range(23,48))
    
    for col in HEADER[24:]:
        train[col].fillna(0, inplace=True)
        train[col] = train[col].astype(int)
    if added_file == 'data/test_ver2.csv':
        added_products = pd.read_csv(added_file, usecols=[1])
    else:
        added_products = pd.read_csv(added_file, usecols=[0])
    
    train.segmento.fillna('02 - PARTICULARES', inplace=True)
    parti = train.loc[train.segmento == '02 - PARTICULARES', :]
    top = train.loc[train.segmento == '01 - TOP', :]
    uni = train.loc[train.segmento == '03 - UNIVERSITARIO', :]
    jacob_pd_p, cosine_pd_p = get_similarity_matrices(parti)
    jacob_pd_t, cosine_pd_t = get_similarity_matrices(top)
    jacob_pd_u, cosine_pd_u = get_similarity_matrices(uni)
    jacobs = {
        '02 - PARTICULARES': (jacob_pd_p, cosine_pd_p),
        '01 - TOP':(jacob_pd_t, cosine_pd_t),
        '03 - UNIVERSITARIO': (jacob_pd_u, cosine_pd_u)
    }
    train.set_index('ncodpers', inplace=True)
    train_final = pd.DataFrame()
    train_final['ncodpers'] = added_products['ncodpers']
    train_final.set_index('ncodpers', inplace=True)
    for col in HEADER[24:]:
        train_final['jacob_custom_' + col] = [0.0]*train_final.shape[0]
        train_final['cosine_custom_' + col] = [0.0]*train_final.shape[0]

    for i in tqdm(train_final.index):
        row = list(train.loc[i, HEADER[24:]])
        segmento = train.loc[i, 'segmento']
        jwts = get_wts_from_mat(jacobs[segmento][0], row)
        cwts = get_wts_from_mat(jacobs[segmento][1], row)
        train_final.loc[i, ['jacob_custom_' + h for h in HEADER[24:]]] = list(jwts)
        train_final.loc[i, ['cosine_custom_' + h for h in HEADER[24:]]] = list(cwts)
    return train_final, file_names[2]

feat, ind = get_similarity_features_custom(('data/train_2015_05_28.csv', 'data/added_product_2015_05_28.csv', 1))

# def get_similarity_features_from_added(file_names): #train_file, added_file):
#     train_file = file_names[0]
#     similarity_file = file_names[1]
#     added_file = file_names[2]
#     print added_file
#     train = pd.read_csv(train_file, header=None, names=HEADER[1:2] + HEADER[24:], usecols=[1]+range(24,48))
#     train.fillna(0, inplace=True)
#     similarity_pd = pd.read_csv(similarity_file)
#     for col in HEADER[24:]:
#         train[col] = train[col].astype(int)
#     if added_file == 'data/test_ver2.csv':
#         added_products = pd.read_csv(added_file, usecols=[1])
#     else:
#         added_products = pd.read_csv(added_file, usecols=[0])
#     jacob_pd, cosine_pd = get_similarity_matrices_from_added(similarity_pd)
#     train.set_index('ncodpers', inplace=True)
#     train_final = pd.DataFrame()
#     train_final['ncodpers'] = added_products['ncodpers']
#     train_final.set_index('ncodpers', inplace=True)
#     for col in HEADER[24:]:
#         train_final['jacob_added_' + col] = [0.0]*train_final.shape[0]
#         train_final['cosine_added_' + col] = [0.0]*train_final.shape[0]

#     for i in tqdm(train_final.index):
#         row = list(train.loc[i, HEADER[24:]])
#         jwts = get_wts_from_mat(jacob_pd, row)
#         cwts = get_wts_from_mat(cosine_pd, row)
#         train_final.loc[i, ['jacob_added_' + h for h in HEADER[24:]]] = list(jwts)
#         train_final.loc[i, ['cosine_added_' + h for h in HEADER[24:]]] = list(cwts)
#     return train_final, jacob_pd, cosine_pd, file_names[3]

file_names_list = [('data/train_2015_05_28.csv', 'data/added_product_2015_05_28.csv', 1),
                   ('data/train_2015_06_28.csv', 'data/added_product_2015_06_28.csv', 2),
                   ('data/train_2015_07_28.csv', 'data/added_product_2015_07_28.csv', 3),
                   ('data/train_2015_08_28.csv', 'data/added_product_2015_08_28.csv', 4),
                   ('data/train_2015_09_28.csv', 'data/added_product_2015_09_28.csv', 5),
                   ('data/train_2015_10_28.csv', 'data/added_product_2015_10_28.csv', 6),
                   ('data/train_2015_11_28.csv', 'data/added_product_2015_11_28.csv', 7),
                   ('data/train_2015_12_28.csv', 'data/added_product_2015_12_28.csv', 8),
                   ('data/train_2016_01_28.csv', 'data/added_product_2016_01_28.csv', 9),
                   ('data/train_2016_02_28.csv', 'data/added_product_2016_02_28.csv', 10),
                   ('data/train_2016_03_28.csv', 'data/added_product_2016_03_28.csv', 11),
                   ('data/train_2016_04_28.csv', 'data/added_product_2016_04_28.csv', 12)
                  ]

test_file_data = ('data/train_2016_05_28.csv', 'data/test_ver2.csv', 13)

p = Pool(5)
results = p.map(get_similarity_features_custom, file_names_list)
p.close()
p.join()

jacob_custom_train = pd.concat(tuple([x[0] for x in results]))

jacob_custom_train.shape

pickle.dump(jacob_custom_train, open('data/15.Jacobian_similarity_features_custom.train_data', 'wb'))
make_submission('data/15.Jacobian_similarity_features_custom.train_data', upload_to_s3=True)

jacob_custom_test, ind = get_similarity_features_custom(test_file_data)
pickle.dump(jacob_custom_test, open('data/15.Jacobian_similarity_features_custom.test_data', 'wb'))
make_submission('data/15.Jacobian_similarity_features_custom.test_data', upload_to_s3=True)

p = Pool(5)
results = p.map(get_similarity_features, file_names_list)
p.close()
p.join()

[x[1] for x in results]

jacob_train = pd.concat(tuple([x[0] for x in results]))

jacob_train.shape

pickle.dump(jacob_train, open('data/15.Jacobian_similarity_features.train_data', 'wb'))
make_submission('data/15.Jacobian_similarity_features.train_data', upload_to_s3=True)

jacob_test, ind = get_similarity_features(test_file_data)

jacob_test.shape

pickle.dump(jacob_test, open('data/15.Jacobian_similarity_features.test_data', 'wb'))
make_submission('data/15.Jacobian_similarity_features.test_data', upload_to_s3=True)

---
# submission from similarty metrics and an actual submission
---

desc, test_ncodpers, preds, label_encoder = pickle.load(open('data/11.xgb_trained_on_all_months_only_added_users_product_features_generated_from_last_4_months.csv.pickle', 'rb'))

len(label_encoder.classes_)

jacobs = pd.DataFrame()
jacobs['ncodpers'] = jacob_test.index
jacobs.set_index('ncodpers', inplace=True)

for col in HEADER[24:]:
    jacobs[col] = jacob_test['jacob_'+col]

preds.shape

jacobs.columns = label_encoder.transform(jacobs.columns)

jacobs.head()

jacobs_1 = jacobs[[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23]]

jacobs_1.head()

mat = jacobs_1.as_matrix()

mat.shape

mat

preds.shape

new_preds = np.multiply(mat, preds)

top_t_products = label_encoder.inverse_transform(np.argsort(new_preds, axis=1)
                                                 [:, ::-1][:, :])

submission = pd.DataFrame()
submission['ncodpers'] = test.index
submission['added_products'] = ['ind_recibo_ult1']*submission.shape[0]
submission.set_index('ncodpers', inplace=True)

test = pd.read_csv('data/test_ver2.csv')

test.fillna(0, inplace=True)

test.set_index('ncodpers', inplace=True)

test_copy = test.loc[test_ncodpers, :]

test_old = pd.read_csv('data/train_2016_05_28.csv', header=None, names=HEADER)

test_old.fillna(0, inplace=True)
test_old.set_index('ncodpers', inplace=True)

test_copy.head()

test_copy['xgb_preds'] = [' '.join(x) for x in top_t_products]
test_copy['added_products'] = ['ind_recibo_ult1']*test_copy.shape[0]
for i in tqdm(test_copy.index):
    zipped = zip(HEADER[24:],
                 test_old.loc[i, 'ind_ahor_fin_ult1':'ind_recibo_ult1'])
    products = [x[0] for x in
                [y for y in zipped if y[1] == 1]]
    pred_products = test_copy.loc[i, 'xgb_preds'].split()
    prod_string = ' '.join(filter(lambda x: x not in products, pred_products))
    test_copy.set_value(i, 'added_products', prod_string)

submission.head()

submission.added_products = test_copy.loc[submission.index, 'added_products']

submission.head()

filename = 'data/15.11_submission_weighted_with_jacobian_wts.csv'
description = '15. 11. submission weighted with jacobian weights'
submission.to_csv(filename, columns=['added_products'])
make_submission(filename, description=description, submit=True, compress=True, upload_to_s3=True)

---
# get similarity features from added products
---

file_names_list_added = [('data/train_2015_05_28.csv', 'data/added_product_2015_04_28.csv', 'data/added_product_2015_05_28.csv', 1),
                   ('data/train_2015_06_28.csv', 'data/added_product_2015_05_28.csv', 'data/added_product_2015_06_28.csv', 2),
                   ('data/train_2015_07_28.csv', 'data/added_product_2015_06_28.csv', 'data/added_product_2015_07_28.csv', 3),
                   ('data/train_2015_08_28.csv', 'data/added_product_2015_07_28.csv', 'data/added_product_2015_08_28.csv', 4),
                   ('data/train_2015_09_28.csv', 'data/added_product_2015_08_28.csv', 'data/added_product_2015_09_28.csv', 5),
                   ('data/train_2015_10_28.csv', 'data/added_product_2015_09_28.csv', 'data/added_product_2015_10_28.csv', 6),
                   ('data/train_2015_11_28.csv', 'data/added_product_2015_10_28.csv', 'data/added_product_2015_11_28.csv', 7),
                   ('data/train_2015_12_28.csv', 'data/added_product_2015_11_28.csv', 'data/added_product_2015_12_28.csv', 8),
                   ('data/train_2016_01_28.csv', 'data/added_product_2015_12_28.csv', 'data/added_product_2016_01_28.csv', 9),
                   ('data/train_2016_02_28.csv', 'data/added_product_2016_01_28.csv', 'data/added_product_2016_02_28.csv', 10),
                   ('data/train_2016_03_28.csv', 'data/added_product_2016_02_28.csv', 'data/added_product_2016_03_28.csv', 11),
                   ('data/train_2016_04_28.csv', 'data/added_product_2016_03_28.csv', 'data/added_product_2016_04_28.csv', 12)
                  ]

test_file_data_added = ('data/train_2016_05_28.csv', 'data/added_product_2016_04_28.csv', 'data/test_ver2.csv', 13)

feats_added = get_similarity_features_from_added(file_names_list_added[0])

feats_added[2].head()

feats_added[0].head()

simi_feat, ind = get_similarity_features(('data/train_2015_05_28.csv', 'data/added_product_2015_05_28.csv', 1))

simi_feat.head()

def get_wts_from_mat(df, row):
    brow = map(bool, row)
    wts = df.loc[brow, :].apply(np.mean)
    wts[brow] = 0
    return wts

simi_feat1, ind = get_similarity_features(('data/train_2015_05_28.csv', 'data/added_product_2015_05_28.csv', 1))

simi_feat1.head()

---
# Custom Jacobian similarities
---

tr = pd.read_csv('data/train_2015_05_28.csv', header=None, names=[HEADER[1]]+HEADER[23:], usecols=[1]+range(23,48))

tr.columns

parti = tr.loc[tr.segmento == '02 - PARTICULARES', :]
top = tr.loc[tr.segmento == '01 - TOP', :]
uni = tr.loc[tr.segmento == '03 - UNIVERSITARIO', :]

parti.shape

parti = tr.loc[tr.segmento == '02 - PARTICULARES', :]
top = tr.loc[tr.segmento == '01 - TOP', :]
uni = tr.loc[tr.segmento == '03 - UNIVERSITARIO', :]

p_o_f_parti = get_persons_of_feature(parti)
p_o_f_top = get_persons_of_feature(top)
p_o_f_uni = get_persons_of_feature(uni)

parti.segmento.value_counts()

tr.segmento.fillna('02 - PARTICULARES', inplace=True)

jacob_pd_p, cosine_pd_p = get_similarity_matrices(p_o_f_parti)
jacob_pd_t, cosine_pd_t = get_similarity_matrices(p_o_f_top)
jacob_pd_u, cosine_pd_u = get_similarity_matrices(p_o_f_uni)

get_similarity_matrices(tr)
